Control.Effect.Implicit.Base

Description

Base module that defines the basic datatypes for implicit-effects.

Documentation

type Effect eff = Monad eff

A less scary type alias to Monad. We hope that implicit-effects can help Haskell beginners to learn how to better structure effectful programs without first going through all the struggles of understanding the machineries of monad, monad transformers, free monad, extensible effects, etc before they can properly write a "good" effectful program. Instead beginners should be taught from first principles about algebraic effects, and go through something similar to the Eff language tutorial, which does a great job explaning algebraic effects without mentioning the M-word.

class EffFunctor (comp :: (Type -> Type) -> Type) where

An EffFunctor comp is parameterized by an Effect eff1 and can be lifted to another Effect eff2 by providing a natural transformation eff1 ~> eff2 to effmap. A simple example for comp is Return.

Methods

effmap :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> comp eff1 -> comp eff2

Lift a computation comp from Effect eff1 to eff2 through the given natural transformation.

class EffSpec (sig :: Type)

Associated Types

type Operation sig = (ops :: (Type -> Type) -> Type) | ops -> sig

The operation type for ops is consumed by computations to perform effect operations. The type is indexed by an Effect type eff that indicates under which monad can the operations be performed. Operation ops must be an EffFunctor to lift effect operations to work on any lifted monad. This is similar to how lifting works in MonadTrans, except here we are lifting the operations manually instead of through typeclasses.

type CoOperation sig = (coop :: Type -> Type) | coop -> sig

The co-operation type for ops is produced by computations when they perform effect operations under a free monad, and is used by effect interpreters such as CoOpHandler to interpret effect operations from a free monad. 'CoOperation ops' must be a functor so that we can use it as the payload type in free monads such as ChurchMonad or FreeMonad

class EffSpec ops => ImplicitOps ops where

An effect operation must also be an instance of EffOps to give computations access to Operation ops through implicit parameter constraints. The FreeOps class hides the machinery of implicit parameters and make them appear like regular constraints with local scope. The definition for EffOps for most effect operations can typically be derived mechanically. We may look into using template Haskell to generate instances for EffOps in future to reduce the boilerplate.

The law for EffOps is

forall ops . withOps ops captureOps === ops

This means any non-trivial instance for EffOps must somehow make use of implicit parameters for the law to hold.

Associated Types

type OpsConstraint ops (eff :: Type -> Type) = (c :: Constraint) | c -> ops eff

The constraint kind for the effect operation under Effect eff. This is typically an implicit parameter with a unique name, e.g. type OpsConstraint FooEff eff = (?fooOps :: Operation FooEff eff). Note that there is a injective type families condition, and given that implicit parameters have a single namespace, users must come out with naming conventions for their custom effects to avoid name clashing that would result in compile-time injectivity violation error.

Methods

withOps :: forall eff r. Effect eff => Operation ops eff -> (OpsConstraint ops eff => r) -> r

Takes an effect operation Operation ops eff and bind it to the implicit parameter specified in OpsConstraint ops eff for the continuation r. The expression r can then use the effect operations without having to explicitly pass them around as function arguments. For the example FooEff, the body for withOps can be defined as withOps fooOps cont = let ?fooOps = fooOps in cont .

captureOps :: forall eff. (Effect eff, OpsConstraint ops eff) => Operation ops eff

If an implicit parameter for the effect operation is available in the context, capture it and return the operation as a value. For the example FooEff, the body for captureOps can be defined as captureOps = ?fooOps.

class (EffSpec ops, Functor (CoOperation ops), EffFunctor (Operation ops)) => FreeOps (ops :: Type) where

A FreeOps ops has associated types that can be used for effect operations and interpretations. We typically use dummy datatypes with empty declaration for ops to signify that ops is only used for tagging effects and are not used at the value level.

Methods

mkFreeOps :: forall t eff. (Effect eff, Effect (t eff)) => (forall a. CoOperation ops a -> t eff a) -> Operation ops (t eff)

The free ops constructor is used for generating free operations that can work under any free monad transformer t and any Effect eff. It is given a payload lifter that can lift a CoOperation ops into the free monad transformed effect t eff, which can be used to construct a free Operation ops under the effect t eff. mkFreeOps is used by FreeEff instances so that they can create a free operation for any free monad.

class (ImplicitOps ops, FreeOps ops) => EffOps ops

type EffConstraint ops eff = (Effect eff, OpsConstraint ops eff)

This is a type alias for the implicit parameter constraint for ops, in addition to requiring eff to be an Effect. This helps reducing boilerplate in generic computations so that we do not have to keep repeating the (Effect eff) constraint in our type signatures.

data NoEff

The trivial effect NoEff have singleton operation and can be trivially satisfied for all Effect eff. It is the identity for the Control.Effect.Implicit.Base.Union type operator such that:

forall ops . ops ∪ NoEff ~= NoEff ∪ ops ~= ops

which by ops1 ~= ops2 means ops1 ⊇ ops2 and ops2 ⊇ ops1, i.e. they satisfies equivalent set of OpsConstraints up to casting equivalents by OpsCast.

data Union ops1 ops2 infixr 7

Combines two effect operations into a new effect operation. Union can be used multiple times to combine multiple effect operations into one.

data LiftEff (eff1 :: Type -> Type) (eff2 :: Type -> Type)

An opaque object-ish effect lifter that can apply effmap to an EffFunctor. We define dedicated datatype instead of using the natural transformation (~>) so that in the case of identity, we can construct a special idLift that skip calling effmap id and return the original EffFunctor. This may save a bit on creating redundant Operation and CoOperation objects during effmap id, particularly when they are used together with Computation.

type (∪) = Union infixr 7

Right associative type operator alias for Union, e.g.

forall ops1 ops2 ops3 :: Type .
  ops1 ∪ ops2 ∪ ops3 === Union ops1 (Union ops2 ops3)

data NoOp (eff :: Type -> Type)

Operation NoEff eff is really just () for all Effect eff. We instead define NoOp with phantom type eff so that the injectivity condition for Operation can be satisfied.

Constructors

NoOp

data NoCoOp r

CoOperation NoEff r is really just () for all return type r. We instead define NoCoOp with phantom type r so that the injectivity condition for CoOperation can be satisfied.

Constructors

NoCoOp

class NoConstraint (eff :: Type -> Type)

OpsConstraint NoEff eff is just the empty constraint () for all Effect eff. We instead define the empty class NoConstraint with trivial instance for all Effect eff so that the injectivity condition for OpsConstraint can be satisfied.

data UnionOps ops1 ops2 (eff :: Type -> Type) infixr 7

The Operation of ops1 ∪ ops2 undef effect eff is the product of the underlying Operation ops1 eff and Operation ops2 eff.

Constructors

UnionOps (ops1 eff) (ops2 eff) infixr 7

(∪) :: forall ops1 ops2 eff. ops1 eff -> ops2 eff -> UnionOps ops1 ops2 eff infixr 7

Right associative term operator alias for UnionOps, e.g.

forall ops1 ops2 ops3 .
  ops1 ∪ ops2 ∪ ops3 === UnionOps ops1 (UnionOps ops2 ops3)

data UnionCoOp ops1 ops2 r

The CoOperation of ops1 ∪ ops2 under return type r is the sum of the underlying CoOperation ops1 r and CoOperation ops2 r.

Constructors

LeftCoOp (ops1 r)

RightCoOp (ops2 r)

leftOps :: forall ops1 ops2 eff. UnionOps ops1 ops2 eff -> ops1 eff

Get the left operation of the Operation (ops1 ∪ ops2) product.

rightOps :: forall ops1 ops2 eff. UnionOps ops1 ops2 eff -> ops2 eff

Get the right operation of the Operation (ops1 ∪ ops2) product.

idLift :: forall eff. Effect eff => LiftEff eff eff

An optimized version of LiftEff when there is no effect lifting required, e.g. when we want to execute a Computation with its original effect.

mkLiftEff :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> LiftEff eff1 eff2

Create a LiftEff from a natural transformation eff1 ~> eff2. This assumes eff1 and eff2 are different, as otherwise we can use idLift for optimized version of LiftEff.

liftEff :: LiftEff eff1 eff2 -> forall x. eff1 x -> eff2 x

The base natural transformation that can be extracted

applyEffmap :: LiftEff eff1 eff2 -> forall comp. EffFunctor comp => comp eff1 -> comp eff2

When applying effect lifting to an EffFunctor, applyEffmap should be used instead of calling effmap directly. This could potentially save some unnecesssary objects recreation in the case of idLift.

joinLift :: LiftEff eff1 eff2 -> forall eff3. Effect eff3 => LiftEff eff2 eff3 -> LiftEff eff1 eff3

An optimizable composition of two natural transformations. This makes sure the no new LiftEff is created when either side is idLift. The behavior for joinLift is as follow:

  joinLift idLift idLift === idLift
  joinLift idLift lift2 === lift2
  joinLift lift1 idLift === lift1
  joinLift lift1 lift2 === mkLiftEff $ (liftEff lift2) . (liftEff lift1)