Control.Effect.Implicit.Base.LiftEff

Documentation

data LiftEff (eff1 :: Type -> Type) (eff2 :: Type -> Type)

An opaque object-ish effect lifter that can apply effmap to an EffFunctor. We define dedicated datatype instead of using the natural transformation (~>) so that in the case of identity, we can construct a special idLift that skip calling effmap id and return the original EffFunctor. This may save a bit on creating redundant Operation and CoOperation objects during effmap id, particularly when they are used together with Computation.

idLift :: forall eff. Effect eff => LiftEff eff eff

An optimized version of LiftEff when there is no effect lifting required, e.g. when we want to execute a Computation with its original effect.

mkLiftEff :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> LiftEff eff1 eff2

Create a LiftEff from a natural transformation eff1 ~> eff2. This assumes eff1 and eff2 are different, as otherwise we can use idLift for optimized version of LiftEff.

liftEff :: LiftEff eff1 eff2 -> forall x. eff1 x -> eff2 x

The base natural transformation that can be extracted

applyEffmap :: LiftEff eff1 eff2 -> forall comp. EffFunctor comp => comp eff1 -> comp eff2

When applying effect lifting to an EffFunctor, applyEffmap should be used instead of calling effmap directly. This could potentially save some unnecesssary objects recreation in the case of idLift.

joinLift :: LiftEff eff1 eff2 -> forall eff3. Effect eff3 => LiftEff eff2 eff3 -> LiftEff eff1 eff3

An optimizable composition of two natural transformations. This makes sure the no new LiftEff is created when either side is idLift. The behavior for joinLift is as follow:

  joinLift idLift idLift === idLift
  joinLift idLift lift2 === lift2
  joinLift lift1 idLift === lift1
  joinLift lift1 lift2 === mkLiftEff $ (liftEff lift2) . (liftEff lift1)