Control.Effect.Implicit.Base.NoEff

Documentation

data NoEff

The trivial effect NoEff have singleton operation and can be trivially satisfied for all Effect eff. It is the identity for the Control.Effect.Implicit.Base.Union type operator such that:

forall ops . ops ∪ NoEff ~= NoEff ∪ ops ~= ops

which by ops1 ~= ops2 means ops1 ⊇ ops2 and ops2 ⊇ ops1, i.e. they satisfies equivalent set of OpsConstraints up to casting equivalents by OpsCast.

data NoOp (eff :: Type -> Type)

Operation NoEff eff is really just () for all Effect eff. We instead define NoOp with phantom type eff so that the injectivity condition for Operation can be satisfied.

Constructors

NoOp

data NoCoOp r

CoOperation NoEff r is really just () for all return type r. We instead define NoCoOp with phantom type r so that the injectivity condition for CoOperation can be satisfied.

Constructors

NoCoOp

class NoConstraint (eff :: Type -> Type)

OpsConstraint NoEff eff is just the empty constraint () for all Effect eff. We instead define the empty class NoConstraint with trivial instance for all Effect eff so that the injectivity condition for OpsConstraint can be satisfied.