Control.Effect.Implicit.Base.Spec

Documentation

class EffOps sig

An effect specification is consist of a dummy datatype sig, an operation type and a co-operation type associated with the sig type. The class EffOps represents the effect operation with an injective type family 'Operation.

Associated Types

type Operation sig = (ops :: (Type -> Type) -> Type) | ops -> sig

The operation type for sig is consumed by computations to perform effect operations. The type is indexed by an Effect type eff that indicates under which monad can the operations be performed.

Operation sig must be an EffFunctor to lift effect operations to work on any lifted monad. This is similar to how lifting works in MonadTrans, except here we are lifting the operations manually instead of through typeclasses.

class EffCoOp sig

An effect specification is consist of a dummy datatype sig, an operation type and a co-operation type associated with the sig type. The class EffCoOp represents the effect co-operation with an injective type family 'CoOperation.

Associated Types

type CoOperation sig = (coop :: Type -> Type) | coop -> sig

The co-operation type for sig is produced by computations when they perform effect operations under a free monad, and is used by effect interpreters such as CoOpHandler to interpret effect operations from a free monad.