Control.Effect.Implicit.Base.Union

Documentation

data Union ops1 ops2 infixr 7

Combines two effect operations into a new effect operation. Union can be used multiple times to combine multiple effect operations into one.

(∪) :: forall ops1 ops2 eff. ops1 eff -> ops2 eff -> UnionOps ops1 ops2 eff infixr 7

Right associative term operator alias for UnionOps, e.g.

forall ops1 ops2 ops3 .
  ops1 ∪ ops2 ∪ ops3 === UnionOps ops1 (UnionOps ops2 ops3)

data UnionOps ops1 ops2 (eff :: Type -> Type) infixr 7

The Operation of ops1 ∪ ops2 undef effect eff is the product of the underlying Operation ops1 eff and Operation ops2 eff.

Constructors

UnionOps (ops1 eff) (ops2 eff) infixr 7

type (∪) = Union infixr 7

Right associative type operator alias for Union, e.g.

forall ops1 ops2 ops3 :: Type .
  ops1 ∪ ops2 ∪ ops3 === Union ops1 (Union ops2 ops3)

data UnionCoOp ops1 ops2 r

The CoOperation of ops1 ∪ ops2 under return type r is the sum of the underlying CoOperation ops1 r and CoOperation ops2 r.

Constructors

LeftCoOp (ops1 r)

RightCoOp (ops2 r)

leftOps :: forall ops1 ops2 eff. UnionOps ops1 ops2 eff -> ops1 eff

Get the left operation of the Operation (ops1 ∪ ops2) product.

rightOps :: forall ops1 ops2 eff. UnionOps ops1 ops2 eff -> ops2 eff

Get the right operation of the Operation (ops1 ∪ ops2) product.