Control.Effect.Implicit.Free

Documentation

class EffCoOp sig

An effect specification is consist of a dummy datatype sig, an operation type and a co-operation type associated with the sig type. The class EffCoOp represents the effect co-operation with an injective type family 'CoOperation.

Associated Types

type CoOperation sig = (coop :: Type -> Type) | coop -> sig

The co-operation type for sig is produced by computations when they perform effect operations under a free monad, and is used by effect interpreters such as CoOpHandler to interpret effect operations from a free monad.

class (EffOps ops, EffCoOp ops, Functor (CoOperation ops), EffFunctor (Operation ops)) => FreeOps (ops :: Type) where

A FreeOps ops has associated types that can be used for effect operations and interpretations. We typically use dummy datatypes with empty declaration for ops to signify that ops is only used for tagging effects and are not used at the value level.

Methods

mkFreeOps :: forall t eff. (Effect eff, Effect (t eff)) => (forall a. CoOperation ops a -> t eff a) -> Operation ops (t eff)

The free ops constructor is used for generating free operations that can work under any free monad transformer t and any Effect eff. It is given a payload lifter that can lift a CoOperation ops into the free monad transformed effect t eff, which can be used to construct a free Operation ops under the effect t eff. mkFreeOps is used by FreeEff instances so that they can create a free operation for any free monad.

data CoOpHandler handler a r eff

Constructors

CoOpHandler

Fields

returnHandler :: a -> eff r

coOpHandler :: CoOperation handler (eff r) -> eff r

class (forall ops eff. (FreeOps ops, Effect eff) => Monad (free ops eff)) => FreeEff free where

Methods

freeOps :: forall ops eff. (FreeOps ops, Effect eff) => Operation ops (free ops eff)

liftFree :: forall ops eff a. (FreeOps ops, Effect eff) => eff a -> free ops eff a

handleFree :: forall ops eff a r. (Effect eff, FreeOps ops) => CoOpHandler ops a r eff -> free ops eff a -> eff r

data NoCoOp r

CoOperation NoEff r is really just () for all return type r. We instead define NoCoOp with phantom type r so that the injectivity condition for CoOperation can be satisfied.

Constructors

NoCoOp

data UnionCoOp ops1 ops2 r

The CoOperation of ops1 ∪ ops2 under return type r is the sum of the underlying CoOperation ops1 r and CoOperation ops2 r.

Constructors

LeftCoOp (ops1 r)

RightCoOp (ops2 r)

newtype ChurchMonad ops eff a

Constructors

ChurchMonad

Fields

runChurchMonad :: forall r. CoOpHandler ops a r eff -> eff r

newtype FreeMonad ops eff a

Constructors

FreeMonad

Fields

unFreeT :: FreeT (CoOperation ops) eff a

newtype GenericCoOpHandler handler eff

Constructors

GenericCoOpHandler (forall a. CoOpHandler handler a a eff)

data ContextualHandler w handler eff

Constructors

ContextualHandler

Fields

runContextualHandler :: forall a. CoOpHandler handler a (w eff a) eff

extractResult :: forall a. w eff a -> eff a

freeLiftEff :: forall free ops eff. (FreeEff free, FreeOps ops, Effect eff) => LiftEff eff (free ops eff)

withCoOpHandler :: forall free handler eff a r. (Effect eff, FreeOps handler, BaseOps handler, FreeEff free) => CoOpHandler handler a r eff -> (OpsConstraint handler (free handler eff) => free handler eff a) -> eff r

withCoOpHandlerAndOps :: forall free ops handler eff a r. (BaseOps ops, FreeOps handler, BaseOps handler, FreeEff free, EffConstraint ops eff) => CoOpHandler handler a r eff -> ((OpsConstraint handler (free handler eff), OpsConstraint ops (free handler eff)) => free handler eff a) -> eff r

withContextualCoOpHandler :: forall free handler eff a r. (Effect eff, FreeOps handler, BaseOps handler, FreeEff free) => CoOpHandler handler a r eff -> (r -> eff a) -> (OpsConstraint handler (free handler eff) => free handler eff a) -> eff a

coopHandlerToPipeline :: forall free ops1 handler eff1 a b. (Effect eff1, BaseOps ops1, FreeOps handler, BaseOps handler, FreeEff free) => Computation ops1 (CoOpHandler handler a b) eff1 -> Pipeline ops1 handler (Return a) (Return b) eff1 eff1

genericCoOpHandlerToPipeline :: forall free ops1 handler eff1. (Effect eff1, BaseOps ops1, FreeOps handler, BaseOps handler, FreeEff free) => Computation ops1 (GenericCoOpHandler handler) eff1 -> GenericPipeline ops1 handler eff1

contextualHandlerToPipeline :: forall free w ops1 handler eff1. (Effect eff1, BaseOps ops1, FreeOps handler, BaseOps handler, FreeEff free) => Computation ops1 (ContextualHandler w handler) eff1 -> GenericPipeline ops1 handler eff1