Control.Effect.Implicit.Freer

Documentation

class EffCoOp ops

Associated Types

type CoOperation ops = (coop :: Type -> Type) | coop -> ops

class (EffOps ops, EffCoOp ops, EffFunctor (Operation ops)) => FreeOps ops where

Methods

mkFreeOps :: forall eff. Effect eff => (forall a. CoOperation ops a -> eff a) -> Operation ops eff

class (forall ops eff. (FreeOps ops, Effect eff) => Monad (free ops eff)) => FreeEff free where

Methods

freeOps :: forall ops eff. (FreeOps ops, Effect eff) => Operation ops (free ops eff)

liftFree :: forall ops eff a. (FreeOps ops, Effect eff) => eff a -> free ops eff a

handleFree :: forall ops eff a r. (Effect eff, FreeOps ops) => CoOpHandler ops a r eff -> free ops eff a -> eff r

newtype FreerMonad ops eff a

Constructors

FreerMonad

Fields

runFreerMonad :: eff (FreerF ops a (FreerMonad ops eff a))

newtype ChurchMonad ops eff a

Constructors

ChurchMonad

Fields

runChurchMonad :: forall r. CoOpHandler ops a r eff -> eff r

data FreerF ops a b where

Constructors

PureF :: a -> FreerF ops a b

FreeF :: forall ops a b x. CoOperation ops x -> (x -> b) -> FreerF ops a b

data CoOpHandler handler a r eff

Constructors

CoOpHandler

Fields

returnHandler :: a -> eff r

coOpHandler :: forall x. CoOperation handler x -> (x -> eff r) -> eff r

withCoOpHandler :: forall free handler eff a r. (Effect eff, EffOps handler, FreeOps handler, ImplicitOps handler, FreeEff free) => CoOpHandler handler a r eff -> (OpsConstraint handler (free handler eff) => free handler eff a) -> eff r

data NoCoOp r

CoOperation NoOp r is really just () for all return type r. We instead define NoCoOp with phantom type r so that the injectivity condition for CoOperation can be satisfied.

data UnionCoOp coop1 coop2 r

The CoOperation of ops1 ∪ ops2 under return type r is the sum of the underlying CoOperation ops1 r and CoOperation ops2 r.

Constructors

LeftOp (coop1 r)

RightOp (coop2 r)

newtype Nest f g a

Constructors

Nest

Fields

unNest :: f (g a)

data CoOpCont coop r where

Constructors

CoOpCont :: forall coop r x. coop x -> (x -> r) -> CoOpCont coop r

newtype Codensity h a

Constructors

Cod

Fields

runCodensity :: forall x. (a -> h x) -> h x

algCod :: forall h f a. Functor f => (forall x. f (h x) -> h x) -> f (Codensity h a) -> Codensity h a

rollCod :: forall eff f a. Effect eff => eff (Codensity (Nest eff f) a) -> Codensity (Nest eff f) a

codensityOps :: forall ops h. (EffOps ops, EffCoOp ops, FreeOps ops) => (forall x r. CoOperation ops x -> (x -> h r) -> h r) -> Operation ops (Codensity h)

toCodensity :: forall free ops eff f a. (EffCoOp ops, Effect eff, FreeOps ops, FreeEff free) => (forall x. CoOpHandler ops x (f x) eff) -> free ops eff a -> Codensity (Nest eff f) a

toCodensity' :: forall free ops eff f a. (EffCoOp ops, Effect eff, FreeOps ops, FreeEff free) => (forall x r. CoOperation ops x -> (x -> eff (f r)) -> eff (f r)) -> free ops eff a -> Codensity (Nest eff f) a