Control.Effect.Implicit.Higher

Documentation

class EffOps sig

Associated Types

type Operation sig = (ops :: (Type -> Type) -> (Type -> Type) -> Type) | ops -> sig

newtype HigherOps ops (eff1 :: Type -> Type) (eff2 :: Type -> Type)

Constructors

HigherOps

Fields

unHigherOps :: ops eff2

newtype LowerOps ops (eff :: Type -> Type)

Constructors

LowerOps

Fields

unLowerOps :: ops eff eff

class (EffOps ops, EffOps ops, Operation ops ~ HigherOps (Operation ops)) => HigherEffOps ops

class (EffOps ops, EffOps ops, Operation ops ~ LowerOps (Operation ops)) => LowerEffOps ops

type Effect eff = Monad eff

A less scary type alias to Monad. We hope that implicit-effects can help Haskell beginners to learn how to better structure effectful programs without first going through all the struggles of understanding the machineries of monad, monad transformers, free monad, extensible effects, etc before they can properly write a "good" effectful program. Instead beginners should be taught from first principles about algebraic effects, and go through something similar to the Eff language tutorial, which does a great job explaning algebraic effects without mentioning the M-word.

data HigherCoOp coop (f :: Type -> Type) a

Constructors

HigherOp (coop a)

class EffCoOp ops

Associated Types

type CoOperation ops = (coop :: (Type -> Type) -> Type -> Type) | coop -> ops

class CoOpFunctor coop where

Methods

liftCoOp :: forall f1 f2 a. (Functor f1, Functor f2) => (forall x. f1 x -> f2 x) -> coop f1 a -> coop f2 a

class (EffCoOp ops, EffCoOp ops, CoOperation ops ~ HigherCoOp (CoOperation ops)) => HigherEffCoOp ops

newtype ContraLift eff1 eff2

Constructors

ContraLift

Fields

runContraLift :: forall a. (forall w. Functor w => (forall x. eff2 x -> eff1 (w x)) -> eff1 (w a)) -> eff2 a

type ContraFree eff f = forall a. (forall w. Functor w => (forall x. f (eff x) -> eff (w x)) -> eff (w (eff (f a)))) -> eff (f a)

identityContraLift :: forall eff. Effect eff => ContraLift eff eff

joinContraLift :: forall eff1 eff2 eff3. (Effect eff1, Effect eff2, Effect eff3) => ContraLift eff1 eff2 -> ContraLift eff2 eff3 -> ContraLift eff1 eff3

class HigherEffFunctor ops where

Methods

invEffmap :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> ContraLift eff1 eff2 -> ops eff1 eff1 -> ops eff2 eff2

class EffFunctor (comp :: (Type -> Type) -> Type) where

An EffFunctor comp is parameterized by an Effect eff1 and can be lifted to another Effect eff2 by providing a natural transformation eff1 ~> eff2 to effmap. A simple example for comp is Return.

Methods

effmap :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> comp eff1 -> comp eff2

Lift a computation comp from Effect eff1 to eff2 through the given natural transformation.

class NoConstraint (eff1 :: Type -> Type) (eff2 :: Type -> Type)

data HUnionOps ops1 ops2 (eff1 :: Type -> Type) (eff2 :: Type -> Type)

Constructors

HUnionOps (ops1 eff1 eff2) (ops2 eff1 eff2)

data HUnionCoOp coop1 coop2 (f :: Type -> Type) r

Constructors

LeftOp (coop1 f r)

RightOp (coop2 f r)

data HigherLiftEff (eff1 :: Type -> Type) (eff2 :: Type -> Type)

Constructors

HigherLiftEff

Fields

baseLiftEff :: forall x. eff1 x -> eff2 x

contraLiftEff :: ContraLift eff1 eff2

class HigherEffFunctor' ops where

Methods

invEffmap' :: forall eff1 eff2. (Effect eff1, Effect eff2) => (forall x. eff1 x -> eff2 x) -> ContraLift eff1 eff2 -> ops eff1 -> ops eff2

type HigherComputation = Computation HigherLiftEff

type HigherOpsHandler ops handler = OpsHandler HigherLiftEff ops handler

type HigherPipeline = Pipeline HigherLiftEff

higherToBaseLiftEff :: forall eff1 eff2. (Effect eff1, Effect eff2) => HigherLiftEff eff1 eff2 -> LiftEff eff1 eff2

toHigherComputation :: forall ops comp eff. (EffOps ops, Effect eff) => BaseComputation ops comp eff -> HigherComputation ops comp eff

coopHandlerToPipeline :: forall free ops1 handler eff1 f a. (Functor f, Effect eff1, EffOps ops1, EffOps handler, ImplicitOps handler, LowerEffOps handler, FreeOps handler, FreeHandler free) => HigherComputation ops1 (CoOpHandler handler f) eff1 -> HigherPipeline ops1 handler (Return a) (Return (f a)) eff1 eff1